REFACTOR: img폴더를 assets폴더 내부로 이동 - img폴더를 assets폴더 내부로 이동시키고 img 폴더 내 파일을 사용하는 파일들의 상대경로 수정
