+++ src/assets/img/icn_folderClose.png
