+++ src/assets/img/ico_radio_6_on.png
