+++ src/assets/img/intro_bg_2.png
