+++ src/assets/img/search-icon-03.png
