+++ src/styles/App.css
@@ -236,7 +236,7 @@ option {
 }
 .main-banner-1-img {
   height: 400px;
-  background: url("../img/main_banner_1.svg") right bottom no-repeat;
+  background: url("../assets/img/main_banner_1.svg") right bottom no-repeat;
   width: 100%;
   display: inline-block;
   padding-right: 50px;
@@ -314,7 +314,7 @@ option {
 }
 .main-banner-2-img {
   height: 400px;
-  background: url("../img/main_banner_2.svg") left bottom no-repeat;
+  background: url("../assets/img/main_banner_2.svg") left bottom no-repeat;
   width: 100%;
   display: inline-block;
   padding-left: 50px;
@@ -383,7 +383,7 @@ a:hover {
   width: 100%;
   position: flex;
   margin: 0 auto;
-  background: #5400cf url(../img/footer.svg) no-repeat center bottom;
+  background: #5400cf url(../assets/img/footer.svg) no-repeat center bottom;
   background-size: cover;
   font-family: "Noto Sans KR", "Nanum Gothic", "나눔고딕", "Noto Sans KR",
     sans-serif;
@@ -519,10 +519,10 @@ a:hover {
   float: right;
 }
 .pei-1 .img-box {
-  background: url(../img/port_img_01_on.png) no-repeat center center;
+  background: url(../assets/img/port_img_01_on.png) no-repeat center center;
 }
 .pei-2 .img-box {
-  background: url(../img/port_img_02_on.png) no-repeat center center;
+  background: url(../assets/img/port_img_02_on.png) no-repeat center center;
 }
 .port-enter-item .p-1,
 .port-enter-item .p-2 {
